Rename random.py to avoid conflicts with numpy (#581)
Renames tests.random to tests.rand to resolve numpy naming conflict; closes #572.
