+++ tests/rand.py
